requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
dill==0.3.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.0
fuzzywuzzy==0.18.0
google-api-core==2.0.0
google-auth==2.0.1
google-cloud-core==2.0.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
idna==3.2
iniconfig==1.1.1
itsdangerous==1.1.0
jaro-winkler==2.0.0
Jinja2==2.11.3
mapply==0.1.7
MarkupSafe==2.0.1
multiprocess==0.70.12.2
numpy==1.21.2
orjson==3.6.2
packaging==21.0
pandas==1.3.2
pathos==0.2.8
pluggy==0.13.1
pox==0.3.0
ppft==1.6.6.4
promise==2.3
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
six==1.16.0
starlette==0.14.2
strsimpy==0.2.0
toml==0.10.2
tqdm==4.62.2
typing-extensions==3.10.0.0
ujson==4.1.0
urllib3==1.26.6
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
